# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudentHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      parameters:
        - { name: subjectId, in: query, required: false, schema: { type: integer } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Tests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Test" }
    post:
      operationId: createTest
      tags: [tests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Test" }
  /tests/{id}:
    patch:
      operationId: updateTest
      tags: [tests]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Test" }
    delete:
      operationId: deleteTest
      tags: [tests]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Test:
      type: object
      properties:
        id: { type: integer }
        subjectId: { type: ["integer", "null"] }
        subjectName: { type: ["string", "null"] }
        subjectColor: { type: ["string", "null"] }
        title: { type: string }
        scheduledAt: { type: string, format: date-time }
        deadline: { type: ["string", "null"], format: date-time }
        status: { type: string }
        result: { type: ["string", "null"] }
        link: { type: ["string", "null"] }
        importance: { type: string }
        needsPreparation: { type: boolean }
        preparationPlan: { type: ["string", "null"] }
      required: [id, title, scheduledAt, status, importance, needsPreparation]
    TestInput:
      type: object
      properties:
        subjectId: { type: ["integer", "null"] }
        title: { type: string }
        scheduledAt: { type: string, format: date-time }
        deadline: { type: ["string", "null"], format: date-time }
        status: { type: string }
        result: { type: ["string", "null"] }
        link: { type: ["string", "null"] }
        importance: { type: string }
        needsPreparation: { type: boolean }
        preparationPlan: { type: ["string", "null"] }
      required: [title, scheduledAt]
...