# lib/api-zod/src/generated/types/testInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface TestInput {
  /** @nullable */
  subjectId?: number | null;
  title: string;
  scheduledAt: Date;
  /** @nullable */
  deadline?: Date | null;
  status?: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  link?: string | null;
  importance?: string;
  needsPreparation?: boolean;
  /** @nullable */
  preparationPlan?: string | null;
}